Separate the network byteswap functions into network_endian.h
